More clean up and commenting
