+++ src/components/MyAccount/ListsTab/ListsTab.test.tsx
@@ -84,7 +84,7 @@ describe("ListsTab", () => {
 
     const sortMenu = screen.getByLabelText("Sort by:", { exact: false })
     userEvent.click(sortMenu)
-    await userEvent.click(screen.getByText("List name (A-Z)"))
+    await userEvent.click(screen.getByText("List name (A - Z)"))
 
     expect(global.fetch).toHaveBeenCalledWith(
       `${BASE_URL}/api/account/lists?patronId=${processedPatron.id}&sort=list_name_asc`
