+++ src/utils/listUtils.ts
@@ -1,3 +1,5 @@
+import ListRecord from "../models/ListRecord"
+
 export const LIST_RECORDS_PER_PAGE = 20
 
 /**
@@ -38,7 +40,7 @@ export const generateListSlug = (name: string) => {
 }
 
 export const listResultsHeading = (list, currentPage) => {
-  const totalResults = list?.records.length || 0
+  const totalResults = list?.recordCount
   const resultsStart = (currentPage - 1) * LIST_RECORDS_PER_PAGE + 1
   const resultsEnd = Math.min(currentPage * LIST_RECORDS_PER_PAGE, totalResults)
   return `Displaying ${
@@ -46,4 +48,53 @@ export const listResultsHeading = (list, currentPage) => {
       ? `${resultsStart}-${resultsEnd} of ${totalResults.toLocaleString()}`
       : totalResults.toLocaleString()
   } item${totalResults === 1 ? "" : "s"}`
+}
+
+export const buildListRecords = (
+  bibData: any[],
+  pageRecords: ListRecord[],
+  activeSort: string
+): ListRecord[] => {
+  // Map of the page's list records keyed by their URI
+  const pageRecordsMap = pageRecords.reduce((acc: any, record: any) => {
+    acc[record.uri] = record
+    return acc
+  }, {})
+
+  // Map over the fetched bib data to instantiate ListRecords (preserve Discovery API sort order)
+  const updatedRecords = bibData.map((bib: any) => {
+    const bibResult = bib.result || bib
+    const uri =
+      bibResult.uri || (bibResult["@id"] ? bibResult["@id"].substring(4) : "")
+    const record = pageRecordsMap[uri] || { uri }
+    // Temp ListRecord
+    const updated = new ListRecord(
+      {
+        uri: record.uri,
+        addedToListDate: new Date().toISOString(),
+      } as any,
+      bibResult
+    )
+    // restore the addedDate
+    updated.addedDate = record.addedDate
+    return updated
+  })
+
+  // Append any records that Discovery API didn't return
+  pageRecords.forEach((record) => {
+    if (!updatedRecords.find((r: any) => r.uri === record.uri)) {
+      updatedRecords.push(record)
+    }
+  })
+
+  // Sort directly if added sort, don't use Discovery API bib sort
+  if (activeSort.includes("added_date")) {
+    updatedRecords.sort((a, b) => {
+      const indexA = pageRecords.findIndex((r) => r.uri === a.uri)
+      const indexB = pageRecords.findIndex((r) => r.uri === b.uri)
+      return indexA - indexB
+    })
+  }
+
+  return updatedRecords
 }
